+++ .travis/install.sh
@@ -20,7 +20,7 @@ if [[ "$DARWIN" = true ]]; then
         eval "$(pyenv init -)"
     fi
 
-    case "${TOX_ENV}" in
+    case "${TOXENV}" in
         py26)
             curl -O https://bootstrap.pypa.io/get-pip.py
             sudo python get-pip.py
@@ -44,9 +44,6 @@ if [[ "$DARWIN" = true ]]; then
             pyenv install 3.4.2
             pyenv global 3.4.2
             ;;
-        py3pep8)
-            sudo apt-get install python3.3 python3.3-dev
-            ;;
         pypy)
             brew upgrade pyenv
             pyenv install pypy-2.4.0
@@ -62,7 +59,7 @@ if [[ "$DARWIN" = true ]]; then
 else
     sudo add-apt-repository -y ppa:fkrull/deadsnakes
 
-    if [[ "${TOX_ENV}" == "pypy" ]]; then
+    if [[ "${TOXENV}" == "pypy" ]]; then
         sudo add-apt-repository -y ppa:pypy/ppa
     fi
 
@@ -76,7 +73,7 @@ else
         sudo apt-get install -y --force-yes libssl-dev/lucid
     fi
 
-    case "${TOX_ENV}" in
+    case "${TOXENV}" in
         py26)
             sudo apt-get install python2.6 python2.6-dev
             ;;