+++ src/cryptography/hazmat/backends/openssl/dsa.py
@@ -57,11 +57,7 @@ class _DSAVerificationContext(object):
             len(self._signature), self._public_key._dsa_cdata)
 
         if res != 1:
-            errors = self._backend._consume_errors()
-            assert errors
-            if res == -1:
-                assert errors[0].lib == self._backend._lib.ERR_LIB_ASN1
-
+            self._backend._consume_errors()
             raise InvalidSignature