+++ src/cryptography/hazmat/backends/openssl/ec.py
@@ -43,13 +43,31 @@ def _ec_key_curve_sn(backend, ec_key):
     assert group != backend._ffi.NULL
 
     nid = backend._lib.EC_GROUP_get_curve_name(group)
-    assert nid != backend._lib.NID_undef
+    # The following check is to find EC keys with unnamed curves and raise
+    # an error for now.
+    if nid == backend._lib.NID_undef:
+        raise NotImplementedError(
+            "ECDSA certificates with unnamed curves are unsupported "
+            "at this time"
+        )
 
     curve_name = backend._lib.OBJ_nid2sn(nid)
     assert curve_name != backend._ffi.NULL
 
     sn = backend._ffi.string(curve_name).decode('ascii')
     return sn
+
+
+def _mark_asn1_named_ec_curve(backend, ec_cdata):
+    """
+    Set the named curve flag on the EC_KEY. This causes OpenSSL to
+    serialize EC keys along with their curve OID which makes
+    deserialization easier.
+    """
+
+    backend._lib.EC_KEY_set_asn1_flag(
+        ec_cdata, backend._lib.OPENSSL_EC_NAMED_CURVE
+    )
 
 
 def _sn_to_elliptic_curve(backend, sn):
@@ -132,6 +150,7 @@ class _ECDSAVerificationContext(object):
 class _EllipticCurvePrivateKey(object):
     def __init__(self, backend, ec_key_cdata):
         self._backend = backend
+        _mark_asn1_named_ec_curve(backend, ec_key_cdata)
         self._ec_key = ec_key_cdata
 
         sn = _ec_key_curve_sn(backend, ec_key_cdata)
@@ -184,6 +203,7 @@ class _EllipticCurvePrivateKey(object):
 class _EllipticCurvePublicKey(object):
     def __init__(self, backend, ec_key_cdata):
         self._backend = backend
+        _mark_asn1_named_ec_curve(backend, ec_key_cdata)
         self._ec_key = ec_key_cdata
 
         sn = _ec_key_curve_sn(backend, ec_key_cdata)